+++ CHANGELOG.md
@@ -2,12 +2,14 @@
 
 ## New Features
 
+ - PR #106 Added multi-GPU initialization
  - PR #167 Added value setter to `device_scalar`
 
 ## Improvements
 
 - PR #161 Use `std::atexit` to finalize RMM after Python interpreter shutdown
 - PR #165 Align memory resource allocation sizes to 8-byte
+- PR #169 Explicit stream argument for device_buffer methods
 
 ## Bug Fixes