+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 ## Bug Fixes
 
+- PR #174 Make `device_buffer` default ctor explicit to work around type_dispatcher issue in libcudf.
 - PR #170 Always build librmm and rmm, but conditionally upload based on CUDA / Python version